Adds basic player controller
Adds player scene, camera, movement and player settings.

This includes the basic setup of a player controller with movement and camera functionality.
The player can move around and jump, and the camera follows the player.
Additionally adds basic input remapping in the control options.

This extracts commonly used information and places it inside of the global blackboard.